save new xml
Former-commit-id: b75a22207664c4a1d69702d46355d23af2ee84a7

+++ trunk/GeoImageViewer/src/main/java/org/geoimage/viewer/core/batch/AnalysisParams.java
@@ -9,7 +9,7 @@ public class AnalysisParams {
 		public String shapeFile="";
 		public String outputFolder="";
 		public float enl=1;
-		public double buffer=0.0;
+		public int buffer=0;
 		public String epsg="EPSG:4326";	
 		public Date startDate;
